refactor: lotto 데이터와 연관성이 적은 로직 분리
